+++ packages/xrpl/src/models/common/index.ts
@@ -2,7 +2,7 @@ export type LedgerIndex = number | ('validated' | 'closed' | 'current')
 
 export interface XRP {
   currency: 'XRP'
-  issuer: never
+  issuer?: never
 }
 
 export interface IssuedCurrency {
@@ -147,4 +147,11 @@ export interface AuthAccount {
   AuthAccount: {
     Account: string
   }
+}
+
+export interface XChainBridge {
+  LockingChainDoor: string
+  LockingChainIssue: Currency
+  IssuingChainDoor: string
+  IssuingChainIssue: Currency
 }