+++ packages/xrpl/HISTORY.md
@@ -3,9 +3,9 @@
 Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xrpl-announce) for release announcements. We recommend that xrpl.js (ripple-lib) users stay up-to-date with the latest stable release.
 
 ## Unreleased
-
 ### Added
 * Added `ports` field to `ServerInfoResponse`
+* Support for the XChainBridge amendment.
 
 ### Fixed
 * Fix request model fields related to AMM